Calculate Rmean

calculate mean reflectance from all layers for given pixel.

Usage

calculate_rmean(raster, extent = NULL, ext = NULL, filename = NULL)

Arguments

raster

a terra SpatRaster of normalized capture data.

extent

an extent or SpatVector used to subset SpatRaster. Defaults to the entire SpatRaster.

ext

character, a graphic format extension.

filename

NULL (default) to write automatically into products, provide full path and ext to override.

Value

one layer terra SpatRaster with calculated Rmean values.

See also